rename to TestCallingState, clean up and added documentation
