+++ packages/react-composites/tests/browser/call/TestCallingState.ts
@@ -1,10 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export type CallingState = {
+/**
+ * Calling state passed to set up playwright test as a query argument
+ */
+export type TestCallingState = {
   remoteParticipants: TestRemoteParticipant[];
 };
 
+/**
+ * Remote participant state used in {@link TestCallingState}
+ */
 export type TestRemoteParticipant = {
   displayName: string;
   isMuted?: boolean;
