+++ packages/react-composites/tests/browser/call/app/index.tsx
@@ -20,15 +20,15 @@ import {
 import { IDS } from '../../common/constants';
 import { isMobile, verifyParamExists } from '../../common/testAppUtils';
 import memoizeOne from 'memoize-one';
-import { FontIcon, Icon, IContextualMenuItem, mergeStyles } from '@fluentui/react';
+import { Icon, IContextualMenuItem, mergeStyles } from '@fluentui/react';
 import { fromFlatCommunicationIdentifier } from '@internal/acs-ui-common';
 import { MockCallAdapter } from './mocks/MockCallAdapter';
-import { CallingState } from '../CallingState';
+import { TestCallingState } from '../TestCallingState';
 
 const urlSearchParams = new URLSearchParams(window.location.search);
 const params = Object.fromEntries(urlSearchParams.entries());
 
-const state = JSON.parse(params.state) as CallingState;
+const state = JSON.parse(params.state) as TestCallingState;
 const useFrLocale = Boolean(params.useFrLocale);
 const showCallDescription = Boolean(params.showCallDescription);
 const injectParticipantMenuItems = Boolean(params.injectParticipantMenuItems);
@@ -39,8 +39,10 @@ function App(): JSX.Element {
 
   useEffect(() => {
     const initialize = async (): Promise<void> => {
-      let newAdapter: CallAdapter = new MockCallAdapter(state);
-      if (!newAdapter) {
+      let newAdapter: CallAdapter = undefined;
+      if (state) {
+        newAdapter = new MockCallAdapter(state);
+      } else {
         const displayName = verifyParamExists(params.displayName, 'displayName');
         const token = verifyParamExists(params.token, 'token');
         const groupId = verifyParamExists(params.groupId, 'groupId');
